refactor: functional testing relocation
functional testing runs in a different process
