+++ package.json
@@ -154,9 +154,9 @@
     "prepublish": "in-publish && npm run build:webui && npm run code:build || not-in-publish",
     "flow": "flow",
     "pretest": "npm run code:build",
-    "test": "cross-env NODE_ENV=test BABEL_ENV=test TZ=UTC jest --maxWorkers 2",
-    "test:unit": "cross-env NODE_ENV=test BABEL_ENV=test TZ=UTC jest '(/test/unit.*\\.spec|/test/webui/.*\\.spec)\\.js' --maxWorkers 2",
-    "test:func": "cross-env NODE_ENV=test BABEL_ENV=test jest '(/test/functional.*\\.func)\\.js' --maxWorkers 2",
+    "test": "cross-env npm run test:unit && npm run test:func",
+    "test:unit": "cross-env NODE_ENV=test BABEL_ENV=test TZ=UTC jest --config ./jest.config.unit.js --maxWorkers 2",
+    "test:func": "cross-env NODE_ENV=test BABEL_ENV=test jest --config ./test/jest.config.func.js --testPathPattern ./test/functional/index*",
     "test:e2e": "cross-env BABEL_ENV=testE2E jest --config ./test/e2e/jest.e2e.config.js --maxWorkers 2",
     "test:all": "npm run test && npm run test:e2e",
     "pre:ci": "npm run lint && npm run build:webui",
