+++ test/functional/package/gzip.js
