+++ test/functional/package/scoped.js
