+++ test/functional/plugins/auth.js
