+++ test/functional/tags/addtag.js
