+++ test/functional/tags/preserve_tags.js
