+++ test/functional/tags/tags.js
