+++ test/jest.config.func.js
@@ -1,0 +1,8 @@
+/* eslint comma-dangle: 0 */
+
+module.exports = {
+  name: 'verdaccio-func-jest',
+  verbose: true,
+  collectCoverage: false,
+  testPathPattern: 'test/functional/functional.spec.js'
+};
